refactor: import 도메인 수정
